refactor: 컴포넌트 폴더 내 컴포넌트 jsx 파일 index.jsx로 파일명 변경
