+++ src/components/Layout.jsx
@@ -1,10 +1,8 @@
 import { getUser } from './fetchApi';
 import { useEffect, useState } from 'react';
 import { Outlet } from 'react-router-dom';
-import Gnb from './shared/Gnb/Gnb';
-import Footer from './shared/Footer/Footer';
-import SearchBar from './shared/SearchBar/SearchBar';
-import AddLinkBar from './shared/AddLinkBar/AddLinkBar';
+import Gnb from './shared/Gnb';
+import Footer from './shared/Footer';
 
 // Gnb, AddLinkBar, SearchBar, Footer 갖춘 기본 레이아웃
 function Layout() {
